ft-printf-bonus-part

+++ ft-printf-bonus/variant_decimal.c
@@ -21,7 +21,7 @@ void	variant_1(o_list *flags)
 	print_width(flags);
 }
 
-void	variant_2(o_list *flags, char *s, unsigned int n)
+void	variant_2(o_list *flags, char *s, unsigned long n)
 {
 	print_sign(flags);
 	print_width(flags);
@@ -29,7 +29,7 @@ void	variant_2(o_list *flags, char *s, unsigned int n)
 	print_number(s, n, ft_strlen(s));
 }
 
-void	variant_3(o_list *flags, char *s, unsigned int n)
+void	variant_3(o_list *flags, char *s, unsigned long n)
 {
 	print_width(flags);
 	print_sign(flags);
@@ -37,7 +37,7 @@ void	variant_3(o_list *flags, char *s, unsigned int n)
 	print_number(s, n, ft_strlen(s));
 }
 
-void	variant_4(o_list *flags, char *s, unsigned int n)
+void	variant_4(o_list *flags, char *s, unsigned long n)
 {
 	flags->width = -flags->width;
 	print_sign(flags);
@@ -46,7 +46,7 @@ void	variant_4(o_list *flags, char *s, unsigned int n)
 	print_width(flags);
 }
 
-void	variant_5(o_list *flags, char *s, unsigned int n)
+void	variant_5(o_list *flags, char *s, unsigned long n)
 {
 	print_sign(flags);
 	print_precision(flags);
